PR feedback + updated qc API
