+++ docs/usage/plot/ehrapy.plot.qc_metrics.rst
@@ -1,0 +1,12 @@
+﻿:github_url: ehrapy.plot.qc_metrics
+
+ehrapy.plot.qc\_metrics
+=======================
+
+.. autofunction:: ehrapy.plot.qc_metrics
+
+.. _sphx_glr_backref_ehrapy.plot.qc_metrics:
+
+.. minigallery:: ehrapy.plot.qc_metrics
+   :add-heading: Gallery
+   :heading-level: -
