--- docs/usage/preprocessing/ehrapy.preprocessing.calculate_qc_metrics.rst
@@ -1,12 +1,0 @@
-﻿:github_url: ehrapy.preprocessing.calculate_qc_metrics
-
-ehrapy.preprocessing.calculate\_qc\_metrics
-===========================================
-
-.. autofunction:: ehrapy.preprocessing.calculate_qc_metrics
-
-.. _sphx_glr_backref_ehrapy.preprocessing.calculate_qc_metrics:
-
-.. minigallery:: ehrapy.preprocessing.calculate_qc_metrics
-   :add-heading: Gallery
-   :heading-level: -
