--- docs/usage/preprocessing/ehrapy.preprocessing.display_qc_metrics.rst
@@ -1,12 +1,0 @@
-﻿:github_url: ehrapy.preprocessing.display_qc_metrics
-
-ehrapy.preprocessing.display\_qc\_metrics
-=========================================
-
-.. autofunction:: ehrapy.preprocessing.display_qc_metrics
-
-.. _sphx_glr_backref_ehrapy.preprocessing.display_qc_metrics:
-
-.. minigallery:: ehrapy.preprocessing.display_qc_metrics
-   :add-heading: Gallery
-   :heading-level: -
