+++ docs/usage/preprocessing/ehrapy.preprocessing.qc_metrics.rst
@@ -1,0 +1,12 @@
+﻿:github_url: ehrapy.preprocessing.qc_metrics
+
+ehrapy.preprocessing.qc\_metrics
+================================
+
+.. autofunction:: ehrapy.preprocessing.qc_metrics
+
+.. _sphx_glr_backref_ehrapy.preprocessing.qc_metrics:
+
+.. minigallery:: ehrapy.preprocessing.qc_metrics
+   :add-heading: Gallery
+   :heading-level: -
