+++ ehrapy/plot/__init__.py
@@ -1,2 +1,3 @@
 from ehrapy.plot._plot_util import *  # noqa: E402,F403
 from ehrapy.plot._scanpy_pl_api import *  # noqa: E402,F403
+from ehrapy.plot.ehrapy_plot.plot_qc import qc_metrics
